Fix delimiter in output file of epoch encoder example
Since Excel assumes commas are the delimiter when the file extension is CSV, the use of tabs as the delimiter in the epoch encoder example results in output files that are read incorrectly by Excel.  This could be resolved by either changing the file extension to TSV or by using commas with the extension CSV. This change does the latter.