Created WritableEpochSource.load() method
This change allows epoch=None when creating a WritableEpochSource object. In this case, the new load() method is called to build an empty dictionary containing the appropriate keys and data types.

Like the save() method, load() can be overridden in subclasses of WritableEpochSource to load epoch data from arbitrary sources. CsvEpochSource implements an example of this.