Fix for incorrect time jump after inserting epoch with EpochEncoder
After using a drop-down menu to change an existing epoch's label and then pressing a shortcut key to insert a new one, the time would jump back to the first epoch, rather than forward one step. This was caused by `on_seek_table` triggering when the table was cleared by `refresh_table`. Interacting with a drop-down menu in the data table before this apparently caused row selection to trigger for the first row in an unexpected way, even if the drop-down menu interacted with was not in the first row.