Click rectangle in EpochEncoder to find epoch in data table
Added a new feature that makes it easier to find epochs in the EpochEncoder's data table. When an epoch's rectangle is clicked in the plot, the corresponding row in the data table is automatically selected. More specifically, the label drop-down menu is selected, which allows the up and down arrow keys to be used to change the label quickly. Time is also moved to the start of the epoch.