Unique, unchanging ids for epochs in WritableEpochSource
ids (really, indices) added to epoch rectangles in EpochEncoder with commit e01f1ba were buggy. They assumed epoch ordering remained in sync for both the plot and data table, but this was not always the case when epoch labels are changed. Unchanging, unique identifiers that allow for unambiguous matching of rectangles to data table rows are needed.

With this commit, id is now distinct from index and is used to track epochs even as indices change due to insertion, deletion, and reordering of epochs. This fixes known bugs where rectangle ids were not in sync with table rows. However, I suspect more work is needed to make the reverse mapping (from table rows to epochs) more robust.

The interface to WritableEpochSource was also cleaned up, so that data can be accessed more simply and consistently using new class properties.