Double-click rectangle to show and match region in EpochEncoder
Double-clicking on a rectangle will now also show the range/region if it is hidden, in addition to changing the region to match the rectangle.

Additionally, this commit changes the event handlers for both `RectItem` (rectangles) and `MyViewBox` (background) to use PyQtGraph's `mouseClickEvent` with the `event.double()` flag, instead of Qt's `mousePressEvent` and `mouseDoubleClickEvent`, since this approach handles better the choice of which object should receive the event.