Add drag to zoom time to EpochEncoder, EpochViewer, SpikeTrainViewer
Code for dragging the right mouse button to adjust `xsize` (time zoom) was confined to `TraceViewer`. This commit moves that code to base classes and added signal connections for `EpochEncoder`, `EpochViewer`, and `SpikeTrainViewer`.

Time zoom could be added to `TimeFreqViewer` (see commented out code), but performance is very poor, so this was not added.