Make sure EpochEncoder rectangles do not draw over label text boxes
`EpochEncoder`'s `RectItem` rectangles were getting destroyed and recreated in the same z-layer as the `TextItem` labels, causing them to obscure the labels. This commit increases the z-layer for the labels so that they are not obscured by the rectangles. It also ensures the labels are drawn above the region indicator.