Improve label legibility
For most of the non-default color schemes (i.e., whenever the colors were not bright green), contrast between label text and the label background or plot contents was poor. This commit both darkens and increases the opacity of the label backgrounds, making the labels much easier to read for non-default color schemes.