Reduce xsize zoom-in speed
If right-click was held and the mouse was dragged quickly to the left, the zoom rate tended to be much too fast, such that it was common to accidentally zoom in to the MIN_XSIZE in an instant. This commit reduces the zoom rate so that this uncontrolled zoom should not happen.