Remove "s" suffixes from time and xsize spinboxes
The "s" in each of the navigation spinboxes prevented these numbers from being changed easily. One could, for example, carefully select the numbers preceding the unit and then type in a new value, but if the unit suffix was selected as well, the data validator on the spinbox would prevent the user from making any changes. This commit removes the suffixes entirely so that changing values is easier.