Fix for seeking non-existent video frames beyond end of video
`get_frame` was sometimes called with a `target_frame` greater than the last frame index. This could occur when dragging the navigation slider all the way to the end, or when trying to step beyond the end of the video. This caused `get_frame_absolut_seek` to spend 250 iterations in its `reseek` loop, and the video would freeze for many seconds until this loop completed.

This commit adds a check to `get_frame`: if the `target_frame` exceeds the final frame index, the final frame is fetched instead.