Fix for seeking non-existent video frames before start of video
`get_frame` was sometimes called with a negative `target_frame`. This could occur when the video's `t_start` was positive, or if `video_times` began with positive values. Like the bug fixed by 535ba9a, this caused `get_frame_absolut_seek` to spend 250 iterations in its `reseek` loop, and the video would freeze for many seconds until this loop completed.

This commit adds a check to `get_frame`: if the `target_frame` is negative, the first frame is fetched instead.

The commit also moves the comparison to `last_frame_index` to the beginning for optimization.