Fix TraceViewer drawing incomplete curves when zoomed in
For TraceViewer, after zooming time in such that only a few points are visible, the curves would end before the right edge of the viewer. If the user zoomed in more, the ends of the lines would eventually travel off the left edge of the viewer and cease to be visible. This was caused by `DataGrabber.get_data` grabbing too few points when it needed to grab some extras outside the visible plot range to ensure curves continued across the entire plotted range. This commit fixes this so that the curves continue across the viewer no matter how much the user zooms in time.