Don't adjust offsets in apply_ygain_zoom
The objective of the code removed by this commit was to keep the medians of the signals fixed in place as ygain was zoomed in or out under `by_channel` or `same_for_all` scale modes. However, because the median is only estimated from a subset of samples and also changes from chunk to chunk, this makes ygain zoom in these modes a non-reversible operation, i.e., zooming in and then zooming out does not return you to the same configuration. Instead, the offsets of the signals can drift away from their original locations and even out of view.

This commit changes the behavior of ygain zoom in `by_channel` and `same_for_all` scale modes by making it leave the offsets unchanged. Consequently, signals are scaled around 0 (baseline) instead of their medians, and the zoom is perfectly reversible.