Fix clicking on current selection in EventList and EpochEncoder table
Selecting items in the EventList or EpochEncoder table would jump time to the start of the event/epoch, unless the clicked item was already selected. In the case of lists/tables containing only one item, there was no way to unselect an item, so in these cases it was impossible to use this feature again after the first use.

This commit changes the trigger for jumping from changing the selection to simply clicking on an item. This allows the user to easily jump to the currently selected item if they have scrolled away from it.