Fix clicking on current selection in DataFrameView
Selecting items in the DataFrameView would jump time to the value in the "time" column, unless the clicked item was already selected. In the case of tables containing only one item, there was no way to unselect an item, so in these cases it was impossible to use this feature again after the first use.

This commit changes the trigger for jumping from changing the selection to simply clicking on an item. This allows the user to easily jump to the currently selected item if they have scrolled away from it.