Allow epoch encoder shortcuts to apply labels to selected range
Previously, pressing a shortcut key for the epoch encoder (default: number keys) would always create a new epoch beginning at the current time with a duration determined by the `new_epoch_step` parameter. This commit adds additional functionality. If the range selector is active, pressing a shortcut key will fill the range with a new epoch with the corresponding label. If the range selector is not active, the old behavior using `new_epoch_step` is used.