Fix trace label and baseline placement for RawIOs with non-zero offset
Fixes #58.

Added new `raw_gains` and `raw_offsets` to `TraceViewer_ParamController`, which are distinguished from the adjustable `gains` and `offsets` parameters. For raw signal sources, these values are obtained from the RawIO header. For in-memory signal sources, `raw_gains` = 1 and `raw_offsets` = 0.

New properties `total_gains` and `total_offsets` are added which provide the combined raw and adjustable gains and offsets. In the case of `total_offsets`, this is not simply the sum as you might expect (see detailed comments for explanation).

Parameters to functions that require the total gains and offsets (`get_data` and `on_request_data`) were renamed to avoid confusion.

Distinguishing between raw and adjustable gains and offsets makes it easier to properly place the trace labels and baselines at each trace's zero for both raw and in-memory sources.

Due to this new division of labor, changes made to `compute_rescale` in #56 are effectively reverted.