Add xratio parameter to viewers for controlling vline position
The parameter may vary between 0 and 1 and controls the fraction of the displayed data that is in the "past", preceding the current time. Default: 0.3.

With xratio set to 1, all displayed data precedes the current t, so only "past" data is shown. With xratio set to 0, all displayed data follows the current t, so only "future" data is shown.