Add editable fields and row buttons to epoch encoder table
The start, stop, and duration fields of the epoch encoder table are now editable. The epoch viewer updates after numbers are changed, as do other dependent table entries (e.g., changing a duration updates the stop time too so that they are consistent).

Removed the "Table operations" buttons, which took up a lot of space, and gave the table a stretch layout parameter so it would fill as much of the control panel as possible. The functions of the removed buttons (epoch splitting, duplication, and deletion) are now handled by new, smaller buttons inserted into every row of the table. Another new button was added for jumping to the epoch's start time (old behavior of clicking on the table row to jump to epoch is now removed). All new buttons have new customized icons and tooltips.