Replace buttons and combos in EpochEncoder table with faster elements
Prior to this change, QPushButtons and QComboBoxes in every row of the EpochEncoder's table were destroyed and re-created with every table refresh, which is very expensive.

These have all been replaced with QTableWidgetItems, which are much faster. QIcons are now created just once at initialization instead of once for each row, each refresh. When a label cell is clicked, it is replaced with a combo box, so labels can still be changed in this way.

These changes result in dramatically improved performance in updating the EpochEncoder table when there are many epochs.