Allow standalone app to open multiple files, each in independent windows
`StandAloneViewer` was replaced by `WindowManager`, which handles spawning a new `MainViewer` each time the Open menu action is used. `WindowManager` also cleans up file locks and deallocates memory when a `MainViewer` is closed.

Additionally, when the app is run from the command line without arguments, the `RawNeoOpenDialog` is launched immediately, rather than the user being required to launch it manually with the Open menu action.