Fix `'method' object is not connected` error for TimeFreqViewer
Fixes #124.

pyqtgraph 0.11.0 added the connection and disconnection of a signal to GraphicsLayout.addItem and GraphicsLayout.removeItem, respectively. ephyviewer's create_plot_grid was bypassing the initial connection by calling not GraphicsLayout.addItem, but GraphicsLayout.layout.addItem. When GraphicsLayout.removeItem was later used, the signal disconnection step raised an error because the signal was never connected in the first place.

This change makes sure the correct addItem method is called, so that the signal is initially connected.