Remove claim of Python 3.4 support
PyQt5 is not available for Python 3.4 from PyPI, so tests are failing. Perhaps it could be built, or perhaps PyQt4 or PySide could work, but Python 3.4 is long past end-of-life, so I think just dropping it makes sense.

Python 3.5 has also reached end-of-life, but I suppose it doesn't hurt to keep it around for now if tests are passing...